refactor(block-editor): extract header into focused components (#1256)
Behavior-neutral decomposition of the 700-line BlockEditorHeader (PR 2 of
the block-editor chrome redesign stack). Splits by concern into HeaderTitleRow,
ViewModeRocker, SaveActions, and HeaderKebab under header/, with getHeaderStyles
moved to header/header.styles.ts. The parent keeps the single-row layout and the
inline status/undo-redo/panel controls unchanged.

Lifts the panel-mode listener + pop-out/full-screen dispatchers into a shared
global-state hook (usePanelModeControls) so later PRs don't duplicate the
listener. The docs-panel usePanelMode keeps its own fullscreen self-heal (its
test mocks setMode, so the self-heal must own local state); the self-heal is
deliberately not carried into the editor, which mounts on the fullscreen surface.

Resolves the #1331 review leave-behind: destructures the stable trackLoadedGuide
and narrows the useCallback deps in BlockEditor.tsx so they no longer recreate
every render off the fresh useBackendSaveFlow return literal.

Zero visual change: external BlockEditorHeaderProps interface, all data-testids,
BlockEditorHeader.test.tsx, and the Playwright spec are unchanged and green.
